+++ apps/pdv-backend/src/modules/fiscal/fiscal.module.ts
@@ -7,6 +7,7 @@ import { FISCAL_DOCUMENT_REPOSITORY } from "./application/ports/fiscal-document-
 import { FISCAL_GATEWAY } from "./application/ports/fiscal-gateway.port.js";
 import { GetFiscalStatusUseCase } from "./application/use-cases/get-fiscal-status.use-case.js";
 import { IssueFiscalReceiptManuallyUseCase } from "./application/use-cases/issue-fiscal-receipt-manually.use-case.js";
+import { RetryFiscalDocumentUseCase } from "./application/use-cases/retry-fiscal-document.use-case.js";
 
 @Module({
   imports: [ProvisioningModule],
@@ -14,6 +15,7 @@ import { IssueFiscalReceiptManuallyUseCase } from "./application/use-cases/issue
   providers: [
     GetFiscalStatusUseCase,
     IssueFiscalReceiptManuallyUseCase,
+    RetryFiscalDocumentUseCase,
     { provide: FISCAL_DOCUMENT_REPOSITORY, useClass: PrismaFiscalDocumentRepository },
     { provide: FISCAL_GATEWAY, useClass: HttpFiscalGateway },
   ],